requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
python-pptx>=0.6.21
scipy>=1.7.0 